update BaseEntity to generic. so 'id' can be any type that extends Number(Long, Integer, Short)

+++ pet-clinic-data/src/main/java/org/learning/spring/guruspringpetclinic/services/map/AbstractMapService.java
@@ -29,7 +29,7 @@ public abstract class AbstractMapService<T extends BaseEntity, ID extends Long> 
 
         if (object.getId() == null)
             object.setId(getNextId());
-        map.put(object.getId(), object);
+        map.put((Long)object.getId(), object);
         return object;
     }
